+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## unreleased
+
+### Fixed
+* Querying trips with limit directly from `Query` instance (#19)
+
+
 ## 1.3.2 - 2022-08-30
 
 ### Improvements